requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==3.0
decorator==4.4.2
idna==2.10
multidict==4.7.6
pycparser==2.20
python-telegram-bot==12.8
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
tornado==6.0.4
urllib3==1.25.9
yarl==1.4.2
